Update README.md
This is the firmware of Panasonic VP7723A audio analyzer, code 14038 is the e2prom program, and 14048.HEX corresponds to the internal program of 14048 e2prom. It can be used by burning it with a programmer. This firmware plays a crucial role in repairing VP7723A battery leakage and RAM power loss, as well as repairing audio analyzers for data loss